+++ src/main/groovy/org/folio/inventory/domain/Instance.groovy
@@ -1,48 +1,60 @@
 package org.folio.inventory.domain
 
+import java.util.stream.Collectors
+
 class Instance {
   final String id
   final String title
+  final String source
+  final String instanceTypeId
   final List<Map> identifiers
+  final List<Creator> creators
 
-  def Instance(String title, List<Map> identifiers) {
-    this(null, title, identifiers)
-  }
+  def Instance(
+    String id,
+    String title,
+    List<Map> identifiers,
+    String source,
+    String instanceTypeId,
+    List<Creator> creators) {
 
-  def Instance(String title) {
-    this(null, title, [])
-  }
-
-  def Instance(String id, String title, List<Map> identifiers) {
     this.id = id
     this.title = title
     this.identifiers = identifiers.collect()
+    this.source = source
+    this.instanceTypeId = instanceTypeId
+    this.creators = creators.collect()
   }
 
   def Instance copyWithNewId(String newId) {
-    new Instance(newId, this.title, this.identifiers)
+    new Instance(newId, this.title, this.identifiers, this.source,
+      this.instanceTypeId, new ArrayList<Creator>())
   }
 
-  Instance addIdentifier(Map identifier) {
-    new Instance(id, title, this.identifiers.collect() << identifier)
+  Instance addIdentifier(String identifierTypeId, String value) {
+    def identifier = ['identifierTypeId' : identifierTypeId, 'value' : value]
+
+    new Instance(id, title, this.identifiers.collect() << identifier,
+      this.source, this.instanceTypeId, this.creators)
   }
 
-  Instance addIdentifier(String namespace, String value) {
-    def identifier = ['namespace' : namespace, 'value' : value]
+  Instance addCreator(String creatorTypeId, String name) {
+    new Instance(id, title, this.identifiers.collect(),
+      this.source, this.instanceTypeId,
+      this.creators.collect() << new Creator(creatorTypeId, name))
+  }
 
-    new Instance(id, title, this.identifiers.collect() << identifier)
+  Instance removeIdentifier(String identifierTypeId, String value) {
+    def newIdentifiers = this.identifiers.stream()
+      .filter({ !(it.identifierTypeId == identifierTypeId && it.value == value) })
+      .collect(Collectors.toList())
+
+    new Instance(id, title, newIdentifiers, this.source, this.instanceTypeId,
+      this.creators)
   }
 
   @Override
   public String toString() {
-    println ("Instance ID: ${id}, Title: ${title}")
-  }
-
-  Instance removeIdentifier(String namespace, String value) {
-    def newIdentifiers = this.identifiers.stream()
-      .filter({ !(it.namespace == namespace && it.value == value) })
-      .collect()
-
-    new Instance(id, title, newIdentifiers)
+    return "Instance ID: ${id}, Title: ${title}";
   }
 }