+++ src/main/groovy/org/folio/inventory/domain/ingest/IngestMessageProcessor.groovy
@@ -6,6 +6,7 @@ import io.vertx.core.json.JsonObject
 import org.folio.inventory.common.CollectAll
 import org.folio.inventory.common.MessagingContext
 import org.folio.inventory.common.domain.Failure
+import org.folio.inventory.domain.Creator
 import org.folio.inventory.domain.Instance
 import org.folio.inventory.domain.Item
 import org.folio.inventory.domain.Messages
@@ -47,7 +48,13 @@ class IngestMessageProcessor {
 
     records.stream()
       .map({
-      new Instance(it.title, JsonArrayHelper.toListOfMaps(it.identifiers))
+        def creators = new ArrayList<Creator>()
+
+        creators.add(new Creator(UUID.randomUUID().toString(), "Fake Author"))
+
+        new Instance(UUID.randomUUID().toString(), it.title,
+          JsonArrayHelper.toListOfMaps(it.identifiers), "Local: MODS",
+          UUID.randomUUID().toString(), creators)
     })
     .forEach({ instanceCollection.add(it, allInstances.receive(),
       { Failure failure -> println("Ingest Creation Failed: ${failure.reason}") })